Refactor: Split prospect service into modules
Split the large prospect service file into smaller, more manageable modules to improve maintainability.
